+++ src/types/settings.d.ts
@@ -390,7 +390,7 @@ declare global {
     "ddb-importer.munching-policy-character-classes": number[];
     "ddb-importer.munching-policy-character-subclasses": Record<string, number[]>;
     "ddb-importer.munching-policy-character-class-rules-version": T5eRulesVersion;
-    "ddb-importer.munching-policy-character-species": number[];
+    "ddb-importer.munching-policy-character-species": string[];
     "ddb-importer.munching-policy-character-dont-grab-existing": boolean;
     "ddb-importer.munching-policy-disable-monster-art": boolean;